+++ web/docusaurus.config.js
@@ -22,6 +22,10 @@ const config = {
 	favicon: 'img/favicon.ico',
 	organizationName: 'pantheon-systems', // Usually your GitHub org/user name.
 	projectName: 'decoupled-kit-js', // Usually your repo name.
+	markdown: {
+		mermaid: true,
+	},
+	themes: ['@docusaurus/theme-mermaid'],
 
 	plugins: [
 		// Prevent trying to generate api reference when building on the platform
@@ -78,6 +82,12 @@ const config = {
 	themeConfig:
 		/** @type {import('@docusaurus/preset-classic').ThemeConfig} */
 		({
+			metadata: [
+				{
+					name: 'keywords',
+					content: 'headless, jamstack, decoupled, wordpress, drupal, webops',
+				},
+			],
 			navbar: {
 				title: 'Decoupled Kit',
 				logo: {
@@ -87,7 +97,7 @@ const config = {
 				items: [
 					{
 						type: 'doc',
-						docId: 'intro',
+						docId: 'decoupled-kit-overview',
 						position: 'left',
 						label: 'Docs',
 					},
@@ -175,6 +185,9 @@ const config = {
 			prism: {
 				theme: lightCodeTheme,
 				darkTheme: darkCodeTheme,
+			},
+			mermaid: {
+				theme: { light: 'neutral', dark: 'dark' },
 			},
 		}),
 };