+++ web/src/pages/index.js
@@ -16,7 +16,7 @@ function HomepageHeader() {
 				<div className={styles.buttons}>
 					<Link
 						className="button button--secondary button--lg"
-						to="/docs/intro"
+						to="/docs/decoupled-kit-overview"
 					>
 						⚡️ See the Docs!
 					</Link>
@@ -30,8 +30,9 @@ export default function Home() {
 	const { siteConfig } = useDocusaurusContext();
 	return (
 		<Layout
-			title={`Hello from ${siteConfig.title}`}
-			description="Description will go into a meta tag in <head />"
+			title="Building Extrordinary Websites"
+			description="Utilities for building a modern front end that sources data from a WordPress or Drupal back end."
+
 		>
 			<HomepageHeader />
 			<main>