+++ packages/nextjs-kit/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @pantheon-systems/nextjs-kit
 
+## 1.6.0-canary.1
+
+### Patch Changes
+
+- a529abf: Replaced `vite-plugin-dts` with `@rollup/plugin-typescript` for type
+  bundling
+
+## 1.6.0-canary.0
+
+### Minor Changes
+
+- b5fae1d: Added Preview Ribbon component to be shared inside Next starters
+
 ## 1.5.0
 
 ### Minor Changes