+++ starters/next-wordpress-starter/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @pantheon-systems/next-wordpress-starter
 
+## 0.12.0-canary.1
+
+### Patch Changes
+
+- 9bf361a: Update preview route due singular values for content types.
+- Updated dependencies [6e00a98]
+- Updated dependencies [a529abf]
+  - @pantheon-systems/wordpress-kit@2.11.0-canary.0
+  - @pantheon-systems/nextjs-kit@1.6.0-canary.1
+
+## 0.12.0-canary.0
+
+### Minor Changes
+
+- b5fae1d: Added clear-preview api route, use the preview ribbon from next-kit
+
+### Patch Changes
+
+- Updated dependencies [b5fae1d]
+  - @pantheon-systems/nextjs-kit@1.6.0-canary.0
+
+## 0.11.0
+
+### Minor Changes
+
+- 5493b01: Add Preview API Route
+
+### Patch Changes
+
+- Updated dependencies [a5c4275]
+- Updated dependencies [a5c4275]
+  - @pantheon-systems/wordpress-kit@2.10.0
+
+## 0.11.0-canary.0
+
+### Minor Changes
+
+- 5493b01: Add Preview API Route
+
+### Patch Changes
+
+- Updated dependencies [a5c4275]
+- Updated dependencies [a5c4275]
+  - @pantheon-systems/wordpress-kit@2.10.0-canary.0
+
 ## 0.10.1
 
 ### Patch Changes